modifiy attachment feature

+++ process.php
@@ -12,7 +12,8 @@ if (empty($POST["body"])) {
     echo "Email body can not be empty";
     exit;
 }
-if (isset($_FILES["attachment"]) && empty($_FILES["attachment"]["name"])) {
+$location = "";
+if (isset($_FILES["attachment"]) && !empty($_FILES["attachment"]["name"])) {
     $dir = "attachment/";
     $location =  $dir.$_FILES["attachment"]["name"];
     $temp_file  =  $FILES['pet_image']['tmp_name'];
@@ -24,6 +25,6 @@ if(move_uploaded_file($temp_file, $location)){
 
     $sender->process_data($POST)
     ->send_email();
-// delete file after sending email;
-    unlink($location);
 }
+// delete file after sending email;
+if(file_exists($location)) unlink($location);
